Rename evento-emergencia to evento-emergencia-medica and update related imports
