+++ src/interfaces/index.ts
@@ -48,7 +48,7 @@ export * from "./ubicacion";
 export * from "./usuario";
 export * from "./destinatario-asistencia";
 export * from "./emergencias-medicas";
-export * from "./evento-emergencia";
+export * from "./evento-emergencia-medica";
 export * from "./personal-salud";
 export * from "./hospitales";
